+++ src/message.c
@@ -1376,6 +1376,8 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "cannot move space because the display is in the middle of an animation.\n");
             } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
                 daemon_fail(rsp, "cannot move space because mission-control is active.\n");
+            } else if (result == SPACE_OP_ERROR_SCRIPTING_ADDITION) {
+                daemon_fail(rsp, "cannot move space due to an error with the scripting-addition.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_SWAP)) {
@@ -1390,6 +1392,8 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "cannot swap space because the display is in the middle of an animation.\n");
             } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
                 daemon_fail(rsp, "cannot swap space because mission-control is active.\n");
+            } else if (result == SPACE_OP_ERROR_SCRIPTING_ADDITION) {
+                daemon_fail(rsp, "cannot swap space due to an error with the scripting-addition.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_DISPLAY)) {
@@ -1408,6 +1412,8 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
                 daemon_fail(rsp, "cannot send space to display because it is in the middle of an animation.\n");
             } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
                 daemon_fail(rsp, "cannot send space to display because mission-control is active.\n");
+            } else if (result == SPACE_OP_ERROR_SCRIPTING_ADDITION) {
+                daemon_fail(rsp, "cannot send space to display due to an error with the scripting-addition.\n");
             }
         }
     } else if (token_equals(command, COMMAND_SPACE_CREATE)) {
@@ -1418,6 +1424,8 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
             daemon_fail(rsp, "cannot create space because the display is in the middle of an animation.\n");
         } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
             daemon_fail(rsp, "cannot create space because mission-control is active.\n");
+        } else if (result == SPACE_OP_ERROR_SCRIPTING_ADDITION) {
+            daemon_fail(rsp, "cannot create space due to an error with the scripting-addition.\n");
         }
     } else if (token_equals(command, COMMAND_SPACE_DESTROY)) {
         enum space_op_error result = space_manager_destroy_space(acting_sid);
@@ -1431,6 +1439,8 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
             daemon_fail(rsp, "cannot destroy space because the display is in the middle of an animation.\n");
         } else if (result == SPACE_OP_ERROR_IN_MISSION_CONTROL) {
             daemon_fail(rsp, "cannot destroy space because mission-control is active.\n");
+        } else if (result == SPACE_OP_ERROR_SCRIPTING_ADDITION) {
+            daemon_fail(rsp, "cannot destroy space due to an error with the scripting-addition.\n");
         }
     } else if (token_equals(command, COMMAND_SPACE_BALANCE)) {
         if (!space_manager_balance_space(&g_space_manager, acting_sid)) {